refactor: restructure code, add src folder
